Example cytoscape subnetwork file
This is an example of the visualized subnetwork. The file should be opened using Cytoscape. The subnetwork displayed is a version that contains two edges per gene-gene pair, as outlined in the report. To generate a subnetwork with one edge per gene-gene pair, please follow the instructions in the README to generate the corresponding output file, to import into Cytoscape.